+++ packages/analytics-js-common/src/types/HttpClient.ts
@@ -15,8 +15,12 @@ export type ResponseDetails = {
   options: IXHRRequestOptions;
 };
 
+export type AsyncRequestCallback<T> = (
+  data?: T | string | undefined,
+  details?: ResponseDetails,
+) => void;
 export interface IAsyncRequestConfig<T> extends IRequestConfig {
-  callback?(data?: T | string | undefined, details?: ResponseDetails): void;
+  callback?: AsyncRequestCallback<T>;
 }
 
 export interface IXHRRequestOptions {