+++ packages/analytics-js/__tests__/components/userSessionManager/UserSessionManager.test.ts
@@ -20,6 +20,8 @@ import {
   entriesWithOnlyNoStorage,
   entriesWithStorageOnlyForAnonymousId,
 } from '../../../__fixtures__/fixtures';
+import { server } from '../../../__fixtures__/msw.server';
+import { defaultHttpClient } from '../../../src/services/HttpClient';
 
 jest.mock('@rudderstack/analytics-js-common/utilities/uuId', () => ({
   generateUUID: jest.fn().mockReturnValue('test_uuid'),
@@ -44,11 +46,9 @@ describe('User session manager', () => {
   let userSessionManager: UserSessionManager;
 
   defaultStoreManager.init();
-  const clientDataStoreCookie = defaultStoreManager.getStore('clientDataInCookie') as Store;
-  const clientDataStoreLS = defaultStoreManager.getStore('clientDataInLocalStorage') as Store;
-  const clientDataStoreSession = defaultStoreManager.getStore(
-    'clientDataInSessionStorage',
-  ) as Store;
+  let clientDataStoreCookie;
+  let clientDataStoreLS;
+  let clientDataStoreSession;
 
   const setDataInLocalStorage = (data: any) => {
     Object.entries(data).forEach(([key, value]) => {
@@ -77,13 +77,20 @@ describe('User session manager', () => {
   };
 
   beforeEach(() => {
+    clientDataStoreCookie = defaultStoreManager.getStore('clientDataInCookie') as Store;
+    clientDataStoreLS = defaultStoreManager.getStore('clientDataInLocalStorage') as Store;
+    clientDataStoreSession = defaultStoreManager.getStore('clientDataInSessionStorage') as Store;
+
     clearStorage();
     resetState();
+
+    state.storage.entries.value = entriesWithOnlyCookieStorage;
     userSessionManager = new UserSessionManager(
       defaultErrorHandler,
       defaultLogger,
       defaultPluginsManager,
       defaultStoreManager,
+      defaultHttpClient,
     );
   });
 
@@ -925,7 +932,6 @@ describe('User session manager', () => {
   describe('setUserId', () => {
     it('should set the provided user ID', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      clientDataStoreCookie.set = jest.fn();
       const newUserId = 'new-dummy-user-id';
       userSessionManager.init();
       userSessionManager.setUserId(newUserId);
@@ -956,7 +962,6 @@ describe('User session manager', () => {
   describe('setUserTraits', () => {
     it('should set the provided user traits', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      clientDataStoreCookie.set = jest.fn();
       const newUserTraits = { key1: 'value1', key2: 'value2' };
       userSessionManager.init();
       userSessionManager.setUserTraits(newUserTraits);
@@ -986,13 +991,15 @@ describe('User session manager', () => {
 
   describe('setGroupId', () => {
     it('should set the provided group id', () => {
+      const cookieStoreSetSpy = jest.spyOn(clientDataStoreCookie, 'set');
+
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      clientDataStoreCookie.set = jest.fn();
       const newGroupId = 'new-dummy-group-id';
+
       userSessionManager.init();
       userSessionManager.setGroupId(newGroupId);
       expect(state.session.groupId.value).toBe(newGroupId);
-      expect(clientDataStoreCookie.set).toHaveBeenCalled();
+      expect(cookieStoreSetSpy).toHaveBeenCalled();
     });
 
     it('should reset the value to default value if persistence is not enabled for group id', () => {
@@ -1019,7 +1026,6 @@ describe('User session manager', () => {
   describe('setGroupTraits', () => {
     it('should set the provided group traits', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      clientDataStoreCookie.set = jest.fn();
       const newGroupTraits = { key1: 'value1', key2: 'value2' };
       userSessionManager.init();
       userSessionManager.setGroupTraits(newGroupTraits);
@@ -1054,7 +1060,6 @@ describe('User session manager', () => {
   describe('setInitialReferrer', () => {
     it('should set the provided initial referrer', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      clientDataStoreCookie.set = jest.fn();
       const newReferrer = 'new-dummy-referrer-1';
       userSessionManager.init();
       userSessionManager.setInitialReferrer(newReferrer);
@@ -1085,7 +1090,6 @@ describe('User session manager', () => {
   describe('setInitialReferringDomain', () => {
     it('should set the provided initial referring domain', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      clientDataStoreCookie.set = jest.fn();
       const newReferrer = 'new-dummy-referrer-2';
       userSessionManager.init();
       userSessionManager.setInitialReferringDomain(newReferrer);
@@ -1121,7 +1125,6 @@ describe('User session manager', () => {
   describe('setAuthToken', () => {
     it('should set the provided auth token', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
-      clientDataStoreCookie.set = jest.fn();
       const newAuthToken = 'new-dummy-auth-token';
       userSessionManager.init();
       userSessionManager.setAuthToken(newAuthToken);
@@ -1150,7 +1153,7 @@ describe('User session manager', () => {
   });
 
   describe('refreshSession', () => {
-    it('refreshSession: should return empty object if any type of tracking is not enabled', () => {
+    it('should return empty object if any type of tracking is not enabled', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
       state.session.sessionInfo.value = {};
@@ -1158,7 +1161,10 @@ describe('User session manager', () => {
       expect(state.session.sessionInfo.value).toStrictEqual({});
     });
 
-    it('refreshSession: should return session id and sessionStart when auto tracking is enabled', () => {
+    it('should return session id and sessionStart when auto tracking is enabled', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      userSessionManager.init();
+
       const futureTimestamp = Date.now() + 5000;
       state.session.sessionInfo.value = {
         autoTrack: true,
@@ -1177,7 +1183,8 @@ describe('User session manager', () => {
       });
     });
 
-    it('refreshSession: should return session id and sessionStart when manual tracking is enabled', () => {
+    it('should return session id and sessionStart when manual tracking is enabled', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
       const manualTrackingSessionId = 1029384756;
       userSessionManager.init();
       userSessionManager.start(manualTrackingSessionId);
@@ -1189,7 +1196,8 @@ describe('User session manager', () => {
       });
     });
 
-    it('refreshSession: should generate new session id and sessionStart and return when auto tracking session is expired', () => {
+    it('should generate new session id and sessionStart and return when auto tracking session is expired', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
       const pastTimestamp = Date.now() - 5000;
       state.session.sessionInfo.value = {
@@ -1209,7 +1217,7 @@ describe('User session manager', () => {
       });
     });
 
-    it('refreshSession: should return only session id from the second event of the auto session tracking', () => {
+    it('should return only session id from the second event of the auto session tracking', () => {
       state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
       userSessionManager.refreshSession(); // sessionInfo For First Event
@@ -1217,17 +1225,30 @@ describe('User session manager', () => {
       expect(state.session.sessionInfo.value.sessionStart).toBe(false);
     });
 
-    it('refreshSession: should return only session id from the second event of the manual session tracking', () => {
+    it('should return only session id from the second event of the manual session tracking', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      userSessionManager.init();
       const manualTrackingSessionId = 1029384756;
       userSessionManager.start(manualTrackingSessionId);
       userSessionManager.refreshSession(); // sessionInfo For First Event
       userSessionManager.refreshSession();
       expect(state.session.sessionInfo.value.sessionStart).toBe(false);
+      userSessionManager.end();
+    });
+
+    it('should not set any session info in state if session info is not available in storage', () => {
+      state.storage.entries.value = entriesWithOnlyNoStorage;
+      userSessionManager.init();
+      userSessionManager.refreshSession();
+      expect(state.session.sessionInfo.value).toStrictEqual({});
     });
   });
 
   describe('getSessionId', () => {
-    it('getSessionId: should return session id for active session', () => {
+    it('should return session id for active session', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      userSessionManager.init();
+
       const futureTimestamp = Date.now() + 10000;
       state.session.sessionInfo.value = {
         autoTrack: true,
@@ -1241,7 +1262,10 @@ describe('User session manager', () => {
       expect(sessionId.toString().length).toBeGreaterThan(0);
     });
 
-    it('getSessionId: should return session id for active session', () => {
+    it('should return null for expired session', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      userSessionManager.init();
+
       const pastTimestamp = Date.now() - 5000;
       state.session.sessionInfo.value = {
         autoTrack: true,
@@ -1254,10 +1278,49 @@ describe('User session manager', () => {
       expect(typeof sessionId).toBe('object');
       expect(sessionId).toBe(null);
     });
+
+    it('should return the session ID from storage if it was modified externally', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+
+      // Set session info in storage where the session has expired
+      const pastTimestamp = Date.now() - 5000;
+      state.session.sessionInfo.value = {
+        autoTrack: true,
+        timeout: 10 * 60 * 1000,
+        expiresAt: pastTimestamp,
+        id: 1683613729115,
+        sessionStart: false,
+      };
+
+      userSessionManager.init();
+
+      // Set custom session info in storage directly
+      const customData = {
+        rl_session: {
+          autoTrack: true,
+          timeout: 10 * 60 * 1000,
+          expiresAt: Date.now() + 10000,
+          sessionStart: false,
+          id: 123456789,
+        },
+      };
+      setDataInCookieStorage(customData);
+
+      const sessionId = userSessionManager.getSessionId();
+      expect(sessionId).toBe(customData.rl_session.id);
+    });
+
+    it('should return null if session info is not available in storage', () => {
+      state.storage.entries.value = entriesWithOnlyNoStorage;
+      userSessionManager.init();
+      const sessionId = userSessionManager.getSessionId();
+      expect(sessionId).toBe(null);
+    });
   });
 
-  describe('auto tracking', () => {
-    it('startAutoTracking: should create a new session in case of invalid session', () => {
+  describe('startOrRenewAutoTracking', () => {
+    it('should create a new session if session is expired', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
       state.session.sessionInfo.value = {
         autoTrack: true,
@@ -1266,7 +1329,7 @@ describe('User session manager', () => {
         id: 1683613729115,
         sessionStart: false,
       };
-      userSessionManager.startOrRenewAutoTracking();
+      userSessionManager.startOrRenewAutoTracking(state.session.sessionInfo.value);
       expect(state.session.sessionInfo.value).toEqual({
         autoTrack: true,
         timeout: 10 * 60 * 1000,
@@ -1276,7 +1339,8 @@ describe('User session manager', () => {
       });
     });
 
-    it('startAutoTracking: should not create a new session in case of valid session', () => {
+    it('should not create a new session in case of active session', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
       userSessionManager.init();
       state.session.sessionInfo.value = {
         autoTrack: true,
@@ -1285,7 +1349,7 @@ describe('User session manager', () => {
         id: 1683613729115,
         sessionStart: false,
       };
-      userSessionManager.startOrRenewAutoTracking();
+      userSessionManager.startOrRenewAutoTracking(state.session.sessionInfo.value);
       expect(state.session.sessionInfo.value).toEqual({
         autoTrack: true,
         timeout: 10 * 60 * 1000,
@@ -1371,16 +1435,297 @@ describe('User session manager', () => {
   });
 
   describe('getExternalAnonymousIdByCookieName', () => {
-    it('Should return null if the cookie value does not exists', () => {
+    it('should return null if the cookie value does not exists', () => {
       const externalAnonymousId =
         userSessionManager.getExternalAnonymousIdByCookieName('anonId_cookie');
       expect(externalAnonymousId).toEqual(null);
     });
-    it('Should return the cookie value if exists', () => {
+    it('should return the cookie value if exists', () => {
       document.cookie = 'anonId_cookie=sampleAnonymousId12345';
       const externalAnonymousId =
         userSessionManager.getExternalAnonymousIdByCookieName('anonId_cookie');
       expect(externalAnonymousId).toEqual('sampleAnonymousId12345');
+    });
+  });
+
+  describe('syncValueToStorage', () => {
+    it('should not call setServerSideCookie method in case isEnabledServerSideCookies state option is not set', () => {
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      const setServerSideCookieSpy = jest.spyOn(userSessionManager, 'setServerSideCookie');
+      userSessionManager.syncValueToStorage('anonymousId', 'dummy_anonymousId');
+      expect(setServerSideCookieSpy).not.toHaveBeenCalled();
+    });
+    it('should call setServerSideCookie method in case isEnabledServerSideCookies state option is set to true', done => {
+      state.serverCookies.isEnabledServerSideCookies.value = true;
+      state.storage.entries.value = entriesWithOnlyCookieStorage;
+      state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
+      clientDataStoreCookie.set = jest.fn();
+      const setServerSideCookieSpy = jest.spyOn(userSessionManager, 'setServerSideCookie');
+      userSessionManager.syncValueToStorage('anonymousId', 'dummy_anonymousId');
+      expect(setServerSideCookieSpy).toHaveBeenCalledWith(
+        [{ name: 'rl_anonymous_id', value: 'dummy_anonymousId' }],
+        expect.any(Function),
+        expect.any(Object),
+      );
+      setTimeout(() => {
+        expect(clientDataStoreCookie.set).toHaveBeenCalled();
+        done();
+      }, 1000);
+    });
+    describe('Cookie should be removed from server side', () => {
+      const testCaseData = [null, undefined, '', {}];
+      it.each(testCaseData)('if value is "%s"', cookieValue => {
+        state.serverCookies.isEnabledServerSideCookies.value = true;
+        state.storage.entries.value = entriesWithOnlyCookieStorage;
+        userSessionManager.setServerSideCookie = jest.fn();
+        userSessionManager.syncValueToStorage('anonymousId', cookieValue);
+        expect(userSessionManager.setServerSideCookie).toHaveBeenCalledWith(
+          [{ name: 'rl_anonymous_id', value: '' }],
+          undefined,
+          expect.any(Object),
+        );
+      });
+    });
+  });
+
+  describe('setServerSideCookie', () => {
+    beforeAll(() => {
+      server.listen();
+    });
+
+    afterAll(() => {
+      server.close();
+    });
+    const mockCookieStore = {
+      encrypt: jest.fn(val => `encrypted_${JSON.parse(val)}`),
+      set: jest.fn(),
+      get: jest.fn(() => 'sample_cookie_value_1234'),
+    };
+    const mockCallback = jest.fn();
+    it('should encrypt cookie value and make request to data service', done => {
+      state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
+      const getEncryptedCookieDataSpy = jest.spyOn(userSessionManager, 'getEncryptedCookieData');
+      const makeRequestToSetCookieSpy = jest.spyOn(userSessionManager, 'makeRequestToSetCookie');
+
+      userSessionManager.setServerSideCookie(
+        [{ name: 'key', value: 'sample_cookie_value_1234' }],
+        () => {},
+        mockCookieStore,
+      );
+      expect(getEncryptedCookieDataSpy).toHaveBeenCalledWith(
+        [{ name: 'key', value: 'sample_cookie_value_1234' }],
+        mockCookieStore,
+      );
+      expect(makeRequestToSetCookieSpy).toHaveBeenCalledWith(
+        [{ name: 'key', value: 'encrypted_sample_cookie_value_1234' }],
+        expect.any(Function),
+      );
+      done();
+    });
+    describe('Network request to Data service is successful', () => {
+      it('should validate cookies are set from the server side', done => {
+        state.source.value = { workspaceId: 'sample_workspaceId' };
+        state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
+        state.storage.cookie.value = {
+          maxage: 10 * 60 * 1000, // 10 min
+          path: '/',
+          domain: 'example.com',
+          samesite: 'Lax',
+        };
+        userSessionManager.setServerSideCookie(
+          [{ name: 'key', value: 'sample_cookie_value_1234' }],
+          mockCallback,
+          mockCookieStore,
+        );
+        setTimeout(() => {
+          expect(mockCookieStore.get).toHaveBeenCalledWith('key');
+          expect(mockCookieStore.get()).toBe('sample_cookie_value_1234');
+          expect(defaultLogger.error).not.toHaveBeenCalledWith(
+            'The server failed to set the key cookie. As a fallback, the cookies will be set client side.',
+          );
+          expect(mockCallback).not.toHaveBeenCalled();
+          done();
+        }, 1000);
+      });
+      it('should set cookies from client side if not successfully set from the server side', done => {
+        state.source.value = { workspaceId: 'sample_workspaceId' };
+        state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
+        state.storage.cookie.value = {
+          maxage: 10 * 60 * 1000, // 10 min
+          path: '/',
+          domain: 'example.com',
+          samesite: 'Lax',
+        };
+        userSessionManager.setServerSideCookie(
+          [{ name: 'key', value: 'sample_cookie_value' }],
+          (name, val) => {
+            mockCookieStore.set(name, val);
+          },
+          mockCookieStore,
+        );
+        setTimeout(() => {
+          expect(mockCookieStore.get).toHaveBeenCalledWith('key');
+          expect(defaultLogger.error).toHaveBeenCalledWith(
+            'The server failed to set the key cookie. As a fallback, the cookies will be set client side.',
+          );
+          expect(mockCookieStore.set).toHaveBeenCalledWith('key', 'sample_cookie_value');
+          done();
+        }, 1000);
+      });
+      it('should log error if not successfully removed from the server', done => {
+        state.source.value = { workspaceId: 'sample_workspaceId' };
+        state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
+        state.storage.cookie.value = {
+          maxage: 10 * 60 * 1000, // 10 min
+          path: '/',
+          domain: 'example.com',
+          samesite: 'Lax',
+        };
+        userSessionManager.setServerSideCookie(
+          [{ name: 'key', value: '' }],
+          undefined,
+          mockCookieStore,
+        );
+        setTimeout(() => {
+          expect(mockCookieStore.get).toHaveBeenCalledWith('key');
+          expect(defaultLogger.error).toHaveBeenCalledWith(
+            'The server failed to remove the key cookie.',
+          );
+          done();
+        }, 1000);
+      });
+    });
+
+    it('should set cookie from client side if data service is down', done => {
+      state.source.value = { workspaceId: 'sample_workspaceId' };
+      state.serverCookies.dataServiceUrl.value =
+        'https://dummy.dataplane.host.com/serverDown/rsaRequest';
+      state.storage.cookie.value = {
+        maxage: 10 * 60 * 1000, // 10 min
+        path: '/',
+        domain: 'example.com',
+        samesite: 'Lax',
+      };
+      userSessionManager.setServerSideCookie(
+        [{ name: 'key', value: 'sample_cookie_value_1234' }],
+        mockCallback,
+        mockCookieStore,
+      );
+      setTimeout(() => {
+        expect(mockCallback).toHaveBeenCalled();
+        done();
+      }, 1000);
+    });
+    it('should set cookie from client side if dataServerUrl is invalid', done => {
+      state.source.value = { workspaceId: 'sample_workspaceId' };
+      state.serverCookies.dataServiceUrl.value =
+        'https://dummy.dataplane.host.com/invalidUrl/rsaRequest';
+      state.storage.cookie.value = {
+        maxage: 10 * 60 * 1000, // 10 min
+        path: '/',
+        domain: 'example.com',
+        samesite: 'Lax',
+      };
+      userSessionManager.setServerSideCookie(
+        [{ name: 'key', value: 'sample_cookie_value_1234' }],
+        mockCallback,
+        mockCookieStore,
+      );
+      setTimeout(() => {
+        expect(mockCallback).toHaveBeenCalled();
+        done();
+      }, 1000);
+    });
+    it('should set cookie from client side if any unhandled error ocurred in serServerSideCookie function', () => {
+      state.source.value = { workspaceId: 'sample_workspaceId' };
+      state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
+      userSessionManager.getEncryptedCookieData = jest.fn(() => {
+        throw new Error('test error');
+      });
+      userSessionManager.onError = jest.fn();
+      state.storage.cookie.value = {
+        maxage: 10 * 60 * 1000, // 10 min
+        path: '/',
+        domain: 'example.com',
+        samesite: 'Lax',
+      };
+      userSessionManager.setServerSideCookie(
+        [{ name: 'key', value: 'sample_cookie_value_1234' }],
+        mockCallback,
+        mockCookieStore,
+      );
+      expect(userSessionManager.onError).toHaveBeenCalledTimes(1);
+      expect(userSessionManager.onError).toHaveBeenCalledWith(
+        new Error('test error'),
+        'Failed to set/remove cookies via server. As a fallback, the cookies will be managed client side.',
+      );
+      expect(mockCallback).toHaveBeenCalledWith('key', 'sample_cookie_value_1234');
+    });
+    describe('getEncryptedCookieData', () => {
+      it('cookie value exists', () => {
+        const encryptedData = userSessionManager.getEncryptedCookieData(
+          [{ name: 'key', value: 'sample_cookie_value_1234' }],
+          mockCookieStore,
+        );
+        expect(mockCookieStore.encrypt).toHaveBeenCalled();
+        expect(encryptedData).toStrictEqual([
+          { name: 'key', value: 'encrypted_sample_cookie_value_1234' },
+        ]);
+      });
+      it('cookie value do not exists', () => {
+        const encryptedData = userSessionManager.getEncryptedCookieData(
+          [{ name: 'key', value: '' }],
+          mockCookieStore,
+        );
+        expect(mockCookieStore.encrypt).not.toHaveBeenCalled();
+        expect(encryptedData).toStrictEqual([{ name: 'key', value: '' }]);
+      });
+    });
+    describe('makeRequestToSetCookie', () => {
+      it('should make external request to exposed endpoint', done => {
+        state.serverCookies.dataServiceUrl.value = 'https://dummy.dataplane.host.com/rsaRequest';
+        state.source.value = { workspaceId: 'sample_workspaceId' };
+        state.storage.cookie.value = {
+          maxage: 10 * 60 * 1000, // 10 min
+          path: '/',
+          domain: 'example.com',
+          samesite: 'Lax',
+        };
+        const getAsyncDataSpy = jest.spyOn(defaultHttpClient, 'getAsyncData');
+        userSessionManager.makeRequestToSetCookie(
+          [{ name: 'key', value: 'encrypted_sample_cookie_value_1234' }],
+          () => {},
+        );
+        expect(getAsyncDataSpy).toHaveBeenCalledWith({
+          url: `https://dummy.dataplane.host.com/rsaRequest`,
+          options: {
+            method: 'POST',
+            data: JSON.stringify({
+              reqType: 'setCookies',
+              workspaceId: 'sample_workspaceId',
+              data: {
+                options: {
+                  maxAge: 10 * 60 * 1000,
+                  path: '/',
+                  domain: 'example.com',
+                  sameSite: 'Lax',
+                  secure: undefined,
+                },
+                cookies: [
+                  {
+                    name: 'key',
+                    value: 'encrypted_sample_cookie_value_1234',
+                  },
+                ],
+              },
+            }),
+            sendRawData: true,
+          },
+          isRawResponse: true,
+          callback: expect.any(Function),
+        });
+        done();
+      });
     });
   });
 });