+++ packages/analytics-js/src/components/capabilitiesManager/detection/dom.ts
@@ -24,6 +24,7 @@ const legacyJSEngineRequiredPolyfills: Record<string, () => boolean> = {
   'Object.entries': () => !isFunction(globalThis.Object.entries),
   'Object.values': () => !isFunction(globalThis.Object.values),
   'Object.assign': () => !isFunction(globalThis.Object.assign),
+  'Object.fromEntries': () => !isFunction(globalThis.Object.fromEntries),
   'Element.prototype.dataset': () => !isDatasetAvailable(),
   // Ideally, we should separate the checks for TextEncoder and TextDecoder but
   // the polyfill service serves them under the same feature name, "TextEncoder".
@@ -34,6 +35,7 @@ const legacyJSEngineRequiredPolyfills: Record<string, () => boolean> = {
   'navigator.sendBeacon': () => !isFunction(globalThis.navigator.sendBeacon),
   // Note, the polyfill service serves both ArrayBuffer and Uint8Array under the same feature name, "ArrayBuffer".
   ArrayBuffer: () => !isFunction(globalThis.Uint8Array),
+  Set: () => !isFunction(globalThis.Set),
 };
 
 const isLegacyJSEngine = (): boolean => {