+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.1.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.0.6...@rudderstack/analytics-js-service-worker@3.1.0) (2024-05-24)
+
+
+### Features
+
+* **analytics-js-service-worker:** auto append batch endpoint ([#1718](https://github.com/rudderlabs/rudder-sdk-js/issues/1718)) ([3e853a5](https://github.com/rudderlabs/rudder-sdk-js/commit/3e853a5c8a2f9dcab614ad6d5e6c9a54f3f61a57))
+
+
+### Bug Fixes
+
+* user sessions behavior ([#1708](https://github.com/rudderlabs/rudder-sdk-js/issues/1708)) ([84e7174](https://github.com/rudderlabs/rudder-sdk-js/commit/84e71744612c8345dc22b8cb0c9362d104eb35e9))
+
 ## [3.0.6](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.0.5...@rudderstack/analytics-js-service-worker@3.0.6) (2024-03-21)
 
 ## [3.0.5](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.0.4...@rudderstack/analytics-js-service-worker@3.0.5) (2024-03-01)